--- avaliacoes/star-wars/src/main/java/Saber.java
@@ -1,11 +1,0 @@
-public class Saber extends Weapon {
-
-    public Saber(String descricao, int dano) {
-        super(descricao, dano);
-    }
-
-    public void slash() {
-        System.out.println("O sabre " + this.getDescricao() + " atacou!");
-    }
-
-}
