+++ avaliacoes/star-wars/src/main/java/assets/model/Blast.java
@@ -1,0 +1,12 @@
+package assets.model;
+public class Blast extends Weapon {
+
+    public Blast(String descricao, int dano) {
+        super(descricao, dano);
+    }
+
+    public void shot() {
+        System.out.println("O blaster " +this.getDescricao() + " atirou!");
+    }
+
+}
