+++ avaliacoes/star-wars/src/main/java/assets/model/Force.java
@@ -1,0 +1,13 @@
+package assets.model;
+
+public interface Force {
+ 
+   public void mindControl();
+
+   public void farseeing();
+
+   public void telepath();
+
+   public void levitation();
+   
+}
