+++ avaliacoes/star-wars/src/main/java/assets/model/Pessoa.java
@@ -1,0 +1,53 @@
+package assets.model;
+public class Pessoa {
+
+    private String nome;
+    private String sobrenome;
+    private String sexo;
+
+    public Pessoa() {
+ 
+    }
+
+    public Pessoa(String nome, String sobrenome) {
+        this.nome = nome;
+        this.sobrenome = sobrenome;
+    }
+
+    public Pessoa(String nome, String sobrenome, String sexo) {
+        this.nome = nome;
+        this.sobrenome = sobrenome;
+        this.sexo = sexo;
+    }
+
+        public void imprime() {
+            System.out.println(this.nome);
+            System.out.println(this.sobrenome);
+            System.out.println(this.sexo);
+        }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getSobrenome() {
+        return sobrenome;
+    }
+
+    public void setSobrenome(String sobrenome) {
+        this.sobrenome = sobrenome;
+    }
+
+    public String getSexo() {
+        return sexo;
+    }
+
+    public void setSexo(String sexo) {
+        this.sexo = sexo;
+    }
+
+}
