+++ avaliacoes/star-wars/src/main/java/assets/model/Saber.java
@@ -1,0 +1,12 @@
+package assets.model;
+public class Saber extends Weapon {
+
+    public Saber(String descricao, int dano) {
+        super(descricao, dano);
+    }
+
+    public void slash() {
+        System.out.println("O sabre " + this.getDescricao() + " atacou!");
+    }
+
+}
