--- avaliacoes/star-wars/src/main/java/interfaces/Force.java
@@ -1,13 +1,0 @@
-package interfaces;
-
-public interface Force {
- 
-   public void mindControl();
-
-   public void farseeing();
-
-   public void telepath();
-
-   public void levitation();
-   
-}
